moving `tester` into `v1`
